scylla-housekeeping: An optional config file
This patch adds an optional config file that can be passed from the
command line.

If the config file is specified and does not exist, the script will
terminate.

The only parameter that is currently available is check-version that can
be either true or false.

The ConfigParser module is used to read the config file, that should be
on the form of:
[housekeeping]
check-version: True

Signed-off-by: Amnon Heiman <amnon@scylladb.com>